+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Changed
 
 * Now requires PHP 7.1 or newer.
+* Pipeline requires processor as first argument.
 
 ## 0.3.0 - 2016-10-13